refactor: Player와 Dealer 중복 코드 제거를 위해 Participant 추가
