+++ src/main/java/domain/Player.java
@@ -1,37 +1,12 @@
 package domain;
 
-import java.util.List;
-
-public class Player {
-    private final Name name;
-    private final Cards cards;
+public class Player extends Participant{
 
     public Player(String name) {
-        this.name = new Name(name);
-        cards = new Cards();
+        super(new Name(name), new Cards());
     }
 
-    public void addCard(Card card) {
-        cards.add(card);
-    }
-
-    public Card getCard(int index) {
-        return cards.getCard(index);
-    }
-
-    public boolean isOverBlackJack() {
-        return cards.getSum() >= BlackjackGame.BLACK_JACK;
-    }
-
-    public Name getName() {
-        return name;
-    }
-
-    public List<String> getCards() {
-        return cards.getCards();
-    }
-
-    public int getCardsSum() {
-        return cards.getSum();
+    public boolean isOverPlayerBlackJack() {
+        return super.getCardsSum() >= BlackjackGame.BLACK_JACK;
     }
 }
